requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.13.3
allure-python-commons==2.13.3
atomicwrites==1.4.1
attrs==22.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
colorama==0.4.5
contextvars==2.4
Faker==14.2.1
idna==3.6
immutables==0.19
importlib-metadata==4.8.3
iniconfig==1.1.1
jmespath==0.10.0
loguru==0.7.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.1.2
pytest==7.0.1
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.18
win32-setctime==1.1.0
zipp==3.6.0
